Game Manager Change Complete!
Game Manager - Working:
A variable for every stage GameManager have was declared and initialized it with as false; the getters functions to get the status values was created; all the functions to change of stage was activated, also an function to show the actual status of every stage was added.

UIManager - Working: The GameManager was added as an object to change the status of the stage and properly execute the functions to load data and show the details menu panel. The funcion called ActivateParacite was name changed to DesactivateDetailsMenu.

Paracite Scene - Working:
The GameManager still stay in the scene, cause it will activate all values we need to execute the events and interactions. All references of GameManager was set  again and all the interactions works. Some interactions are still needed to implement (See Pending Tasks).

MenuGameScene / btns - Working:
The function to change of stage in OnClick setting was removed to be handle only by the scene that owns it; the event to switch between scenes stays.

Pending Tasks:
1. Models instantiate in plane correction.
2. Delete models when they are update.
3. Add random model selection in GameScene.