Spawn in AR Scene Works!
ARInteractionsManager - Working:
The most important changes was made here, actually it's the script to generate models and interact with them in the AR scene, the function to randomly pick the model was defined in a function called SpawnParasite, to have a better management of the models I add some conditionals in the Update function, this includes the activation of the button to catch.

GameScene - Modified:
Really to many changes was made here, specially many objects to have a good interaction.

Parasite Scene - Edited:
The scroll view modified in the interaction animation.

UIManager - Edited:
Deleted thrash code.

Pending Tasks:
1. Models instantiate in plane revision (the scale is working so bad and I cant to check the if the model spawn in the plane generated by the points cloud).
2. Show all description info in a new panel.
3. Create and add the missing models and the details information.