Detail Panel and Subsystem Creation
- ParaciteScene/MainPanel - Edited:
The layout priorities was changed in the Main Panel for a better UI interaction experience, also the sizes of the buttos of detailsMenuPanel was changed.

- ARInteractionsManager/Random.Range - Edited:
The function picked the same number, so, checking the documentation I realize that the range not include the max number, it was changed.

- Delete Prefabs in Game Scene:
In the ARInteractionsManager this function was fixed, but not directly upon DeleteModel function, but  in the instance of the model, that let me a better modification of the instantiated model.

- Spawn Scale correction:
The change mentioned before.

- ParaciteScene/Detail Panel - Created:
I create a UI Panel for only the detail, all the system to load the detail information automatically but needs revision, probably the issues to generate is for the prefab, 'cause the scriptable objects are right.

Pending Tasks:
1. Models instantiate in plane revision (the scale is working so bad and I cant to check the if the model spawn in the plane generated by the points cloud).
2. Show all description info in a new panel (Created but need revision).
3. Create and add the missing models and the details information.