Details Panel Works!
- ParacitesScene/DetailsPanel - Edited:
The auto generating of the details in the panel was disable to set stay them, and with a reference fill the info.

- DetailContentManager - Edited:
Every single container in the Details Panel has a different size, depending how much information is save. Through the functions expand and refresh spacing the size of this container change and reorganize.

- Character - Edited:
Spelling correction in attribute Symptoms.

- CharacterButtonManager - Edited:
 All the attributes about details of the character was implemented as the others, defining the rout of the text object will load and show the data saved.

- DataManager - Edited:
The details data saved in the scriptable objects is load when the character buttons are created.

- GameManager - Edited:
Status created to handle the details panel.

- UIManager - Edited:
The detail panel stays scaled in 0 and throw an animation to show it. For now the lateral panel (detail menu panel) is disabled, until the detail panel auto focus be implemented.

Pending Tasks:
1. Models instantiate in plane revision (the scale is working so bad and I cant to check the if the model spawn in the plane generated by the points cloud).
2. Create and add the missing models and the details information.
3. Create animations for the 3d models to interact.
3. Implement interactions.